reorg: pull mixed solver wrapper into separate header file
